+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Features:
 - Expose `WorkspaceFunction` definitions.
 - Expose base configuration types.
 - Implement the `ProcessService`.
+- Implement the `GitService`.